created actions/claims, removed file_actions reducer and selector and moved corresponding methods to claims
