--- ui/js/selectors/file_actions.js
@@ -1,13 +1,0 @@
-import { createSelector } from "reselect";
-
-export const _selectState = state => state.fileActions || {};
-
-export const selectClaimSupport = createSelector(
-  _selectState,
-  state => state.claimSupport || {}
-);
-
-export const selectClaimSupportAmount = createSelector(
-  selectClaimSupport,
-  claimSupport => claimSupport.amount
-);
